tests refactoring to the model AAA (arrange, act, assert)
