+++ planet_wttd/util/status_code.py
@@ -1,0 +1,6 @@
+
+
+class Status:
+    status_200_ok = 200
+    status_404_not_found = 404
+    # ToDo add more status_code
